check if file exists

+++ rules/configuration.smk
@@ -259,7 +259,8 @@ def get_layout_trace(sample):
 # multiple writes/reads at the sametime to layouts.p
 
 # let's ignore locks that are older than 5 minutes
-if os.stat(layout_cachefile_lock).st_mtime > 5 * 60:
+if os.exists(layout_cachefile_lock) and \
+        time.time() - os.stat(layout_cachefile_lock).st_mtime > 5 * 60:
     os.remove(layout_cachefile_lock)
 
 with FileLock(layout_cachefile_lock):
